@@ -1,7 +1,7 @@
 /*******************************************************************************
 * Author    :  Angus Johnson                                                   *
 * Version   :  Clipper2 - ver.1.0.4                                            *
-* Date      :  7 September 2022                                                *
+* Date      :  22 September 2022                                               *
 * Website   :  http://www.angusj.com                                           *
 * Copyright :  Angus Johnson 2010-2022                                         *
 * Purpose   :  This is the main polygon clipping module                        *
@@ -216,10 +216,10 @@ namespace Clipper2Lib {
 			b2 = e2.bot.x - e2.bot.y * e2.dx;
 			double q = (b2 - b1) / (e1.dx - e2.dx);
 			return (abs(e1.dx) < abs(e2.dx)) ?
-				Point64(static_cast<int64_t>(std::round(e1.dx * q + b1)),
-					static_cast<int64_t>(std::round(q))) :
-				Point64(static_cast<int64_t>(std::round(e2.dx * q + b2)),
-					static_cast<int64_t>(std::round(q)));
+				Point64(static_cast<int64_t>((e1.dx * q + b1)),
+					static_cast<int64_t>((q))) :
+				Point64(static_cast<int64_t>((e2.dx * q + b2)),
+					static_cast<int64_t>((q)));
 		}
 	}
 
@@ -2820,6 +2820,13 @@ namespace Clipper2Lib {
 	inline bool PointBetween(Point64 pt, Point64 corner1, Point64 corner2)
 	{
 		//NB points may not be collinear
+		return ValueBetween(pt.x, corner1.x, corner2.x) &&
+			ValueBetween(pt.y, corner1.y, corner2.y);
+	}
+
+	inline bool PointEqualOrBetween(Point64 pt, Point64 corner1, Point64 corner2)
+	{
+		//NB points may not be collinear
 		return ValueEqualOrBetween(pt.x, corner1.x, corner2.x) &&
 			ValueEqualOrBetween(pt.y, corner1.y, corner2.y);
 	}
@@ -2960,20 +2967,16 @@ namespace Clipper2Lib {
 	{
 		for (Joiner* j : joiner_list_)
 		{
-			if (!j)
-			{
-				continue;
-			}
-			else if (!succeeded_)
-			{
-				delete j;
-			}
-			else
+			if (!j) continue;
+			if (succeeded_)
 			{
 				OutRec* outrec = ProcessJoin(j);
 				CleanCollinear(outrec);
 			}
+			else 
+				delete j;
 		}
+
 		joiner_list_.resize(0);
 	}
 
@@ -3118,7 +3121,7 @@ namespace Clipper2Lib {
 					//by inserting an extra vertex if needed
 					if (op1->prev->pt != op2->next->pt)
 					{
-						if (PointBetween(op1->prev->pt, op2->pt, op2->next->pt))
+						if (PointEqualOrBetween(op1->prev->pt, op2->pt, op2->next->pt))
 							op2->next = InsertOp(op1->prev->pt, op2);
 						else
 							op1->prev = InsertOp(op2->next->pt, op1->prev);
@@ -3179,7 +3182,7 @@ namespace Clipper2Lib {
 					//by inserting an extra vertex if needed
 					if (op2->prev->pt != op1->next->pt)
 					{
-						if (PointBetween(op2->prev->pt, op1->pt, op1->next->pt))
+						if (PointEqualOrBetween(op2->prev->pt, op1->pt, op1->next->pt))
 							op1->next = InsertOp(op2->prev->pt, op1);
 						else
 							op2->prev = InsertOp(op1->next->pt, op2->prev);
