@@ -1,31 +1,34 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
-using System.Collections.Generic;
 
 namespace Clipper2Lib.UnitTests
 {
-  using Paths64 = List<List<Point64>>;
 
   [TestClass]
   public class TestPolygons
   {
 
+    private bool IsInList(int num, int[] list)
+    {
+      foreach (int i in list) if (i == num) return true;
+      return false;
+    }
+
     [TestMethod]
     public void TestClosedPaths()
     {
-      int i = 0;
+      int testNum = 0;
       while (true)
       {
-        i++;
+        testNum++;
         Clipper64 c64 = new();
         Paths64 subj = new(), subj_open = new (), clip = new();
         Paths64 solution = new(), solution_open = new();
 
         if (!ClipperFileIO.LoadTestNum("..\\..\\..\\..\\..\\..\\Tests\\Polygons.txt",
-          i, subj, subj_open, clip, out ClipType clipType, out FillRule fillrule, 
-          out long area, out int count, out _))
+          testNum, subj, subj_open, clip, out ClipType clipType, out FillRule fillrule, 
+          out long storedArea, out int storedCount, out _))
         {          
-          Assert.IsTrue(i > 180, string.Format("Loading test polygon {0} failed.", i));
+          Assert.IsTrue(testNum > 180, string.Format("Loading test polygon {0} failed.", testNum));
           break;
         }
 
@@ -33,22 +36,40 @@ namespace Clipper2Lib.UnitTests
         c64.AddOpenSubject(subj_open);
         c64.AddClip(clip);
         c64.Execute(clipType, fillrule, solution, solution_open);
+        int measuredCount = solution.Count();
+        long measuredArea = (long)Clipper.Area(solution);
+        int countDiff = storedCount > 0 ? Math.Abs(storedCount - measuredCount) : 0;
+        long areaDiff = storedArea > 0 ? Math.Abs(storedArea - measuredArea) : 0;
 
-        if (area > 0)
+        if (testNum == 23)
         {
-          double area2 = Clipper.Area(solution);
-          double a = area / area2;
-          Assert.IsTrue(Math.Abs(area - area2) < 2 || (a > 0.995 && a < 1.005),
-            string.Format("Incorrect area in test {0}", i));
+          Assert.IsTrue(countDiff <= 4);
         }
-
-        if (count > 0 && Math.Abs(solution.Count - count) > 2 && 
-            (double)Math.Abs(solution.Count - count)/count > 0.02)
+        else if (testNum == 27)
         {
-          Assert.IsTrue(Math.Abs(solution.Count - count) < 4,
-            string.Format("Incorrect count in test {0}", i));
+          Assert.IsTrue(countDiff <= 2);
         }
+        else if (IsInList(testNum,
+          new int[] {18, 32, 42, 43, 45, 87, 102, 103, 111, 118, 183 }))
+        {
+          Assert.IsTrue(countDiff <= 1);
+        }
+        else if (testNum >= 120)
+        {
+          if (storedCount > 0)
+            Assert.IsTrue(countDiff / storedCount <= 0.02);
+        }
+        else if (storedCount > 0)
+          Assert.IsTrue(countDiff == 0);
 
+        if (IsInList(testNum, new int[] { 22,23,24 }))
+        {
+          Assert.IsTrue(areaDiff <= 8);
+        }
+        else if (storedArea > 0 && areaDiff > 100)
+        {
+          Assert.IsTrue(areaDiff / storedArea <= 0.02);
+        }
       } //bottom of num loop
 
     }
