@@ -1,7 +1,7 @@
 ﻿/*******************************************************************************
 * Author    :  Angus Johnson                                                   *
 * Version   :  Clipper2 - ver.1.0.4                                            *
-* Date      :  7 September 2022                                                *
+* Date      :  22 September 2022                                               *
 * Website   :  http://www.angusj.com                                           *
 * Copyright :  Angus Johnson 2010-2022                                         *
 * Purpose   :  This is the main polygon clipping module                        *
@@ -18,11 +18,6 @@ using System.Runtime.CompilerServices;
 
 namespace Clipper2Lib
 {
-
-  using Path64  = List<Point64>;
-  using Paths64 = List<List<Point64>>;
-  using PathD   = List<PointD>;
-  using PathsD  = List<List<PointD>>;
 
   // Vertex: a pre-clipping data structure. It is used to separate polygons
   // into ascending and descending 'bounds' (or sides) that start at local
@@ -173,7 +168,7 @@ namespace Clipper2Lib
 
     public Joiner(OutPt op1, OutPt? op2, Joiner? nextH)
     {
-      this.idx = -1;
+      idx = -1;
       this.nextH = nextH;
       this.op1 = op1;
       this.op2 = op2;
@@ -431,8 +426,10 @@ namespace Clipper2Lib
       b2 = ae2.bot.X - ae2.bot.Y * ae2.dx;
       double q = (b2 - b1) / (ae1.dx - ae2.dx);
       return (Math.Abs(ae1.dx) < Math.Abs(ae2.dx))
-        ? new Point64((long) Math.Round(ae1.dx * q + b1), (long) Math.Round(q))
-        : new Point64((long) Math.Round(ae2.dx * q + b2), (long) Math.Round(q));
+        //? new Point64((long) Math.Round(ae1.dx * q + b1), (long) Math.Round(q))
+        //: new Point64((long) Math.Round(ae2.dx * q + b2), (long) Math.Round(q));
+        ? new Point64((long)(ae1.dx * q + b1), (long)(q))
+        : new Point64((long)(ae2.dx * q + b2), (long)(q));
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -451,7 +448,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private Vertex PrevPrevVertex(Active ae)
+    private static Vertex PrevPrevVertex(Active ae)
     {
       if (ae.windDx > 0)
         return ae.vertexTop!.prev!.prev!;
@@ -471,7 +468,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private Active? GetMaximaPair(Active ae)
+    private static Active? GetMaximaPair(Active ae)
     {
       Active? ae2;
       ae2 = ae.nextInAEL;
@@ -520,9 +517,9 @@ namespace Clipper2Lib
       {
         if (a.pt.Y == b.pt.Y)
         {
+          if (a.pt.X == b.pt.X) return 0;
           return (a.pt.X < b.pt.X) ? -1 : 1;
         }
-
         return (a.pt.Y > b.pt.Y) ? -1 : 1;
       }
     }
@@ -610,7 +607,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private bool OutrecIsAscending(Active hotEdge)
+    private static bool OutrecIsAscending(Active hotEdge)
 	  {
 		  return (hotEdge == hotEdge.outrec!.frontEdge);
 	  }
@@ -1040,7 +1037,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private bool IsValidAelOrder(Active resident, Active newcomer)
+    private static bool IsValidAelOrder(Active resident, Active newcomer)
     {
       if (newcomer.curX != resident.curX)
         return newcomer.curX > resident.curX;
@@ -1111,7 +1108,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private void InsertRightEdge(Active ae, Active ae2)
+    private static void InsertRightEdge(Active ae, Active ae2)
     {
       ae2.nextInAEL = ae.nextInAEL;
       if (ae.nextInAEL != null) ae.nextInAEL.prevInAEL = ae2;
@@ -1265,7 +1262,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private bool TestJoinWithPrev1(Active e)
+    private static bool TestJoinWithPrev1(Active e)
     {
       // this is marginally quicker than TestJoinWithPrev2
       // but can only be used when e.PrevInAEL.currX is accurate
@@ -1276,7 +1273,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private bool TestJoinWithPrev2(Active e, Point64 currPt)
+    private static bool TestJoinWithPrev2(Active e, Point64 currPt)
     {
       return IsHotEdge(e) && !IsOpen(e) &&
              (e.prevInAEL != null) && !IsOpen(e.prevInAEL) &&
@@ -1286,7 +1283,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private bool TestJoinWithNext1(Active e)
+    private static bool TestJoinWithNext1(Active e)
     {
       // this is marginally quicker than TestJoinWithNext2
       // but can only be used when e.NextInAEL.currX is accurate
@@ -1297,7 +1294,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private bool TestJoinWithNext2(Active e, Point64 currPt)
+    private static bool TestJoinWithNext2(Active e, Point64 currPt)
     {
       return IsHotEdge(e) && !IsOpen(e) &&
              (e.nextInAEL != null) && !IsOpen(e.nextInAEL) &&
@@ -1408,7 +1405,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private void JoinOutrecPaths(Active ae1, Active ae2)
+    private static void JoinOutrecPaths(Active ae1, Active ae2)
     {
       // join ae2 outrec path onto ae1 outrec path and then delete ae2 outrec path
       // pointers. (NB Only very rarely do the joining ends share the same coords.)
@@ -1467,7 +1464,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private OutPt AddOutPt(Active ae, Point64 pt)
+    private static OutPt AddOutPt(Active ae, Point64 pt)
     {
       OutPt newOp;
 
@@ -1538,7 +1535,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private Active? FindEdgeWithMatchingLocMin(Active e)
+    private static Active? FindEdgeWithMatchingLocMin(Active e)
     {
       Active? result = e.nextInAEL;
       while (result != null)
@@ -1918,7 +1915,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private Active? ExtractFromSEL(Active ae)
+    private static Active? ExtractFromSEL(Active ae)
     {
       Active? res = ae.nextInSEL;
       if (res != null)
@@ -1928,7 +1925,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private void Insert1Before2InSEL(Active ae1, Active ae2)
+    private static void Insert1Before2InSEL(Active ae1, Active ae2)
     {
       ae1.prevInSEL = ae2.prevInSEL;
       if (ae1.prevInSEL != null)
@@ -2057,7 +2054,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private bool ResetHorzDirection(Active horz, Active? maxPair,
+    private static bool ResetHorzDirection(Active horz, Active? maxPair,
         out long leftX, out long rightX)
     {
       if (horz.bot.X == horz.top.X)
@@ -2082,14 +2079,14 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private bool HorzIsSpike(Active horz)
+    private static bool HorzIsSpike(Active horz)
     {
       Point64 nextPt = NextVertex(horz).pt;
       return (horz.bot.X < horz.top.X) != (horz.top.X < nextPt.X);
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private void TrimHorz(Active horzEdge, bool preserveCollinear)
+    private static void TrimHorz(Active horzEdge, bool preserveCollinear)
     {
       bool wasTrimmed = false;
       Point64 pt = NextVertex(horzEdge).pt;
@@ -2456,12 +2453,20 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static bool PointBetween(Point64 pt, Point64 corner1, Point64 corner2)
+    private static bool PointEqualOrBetween(Point64 pt, Point64 corner1, Point64 corner2)
     {
       // NB points may not be collinear
       return
         ValueEqualOrBetween(pt.X, corner1.X, corner2.X) &&
         ValueEqualOrBetween(pt.Y, corner1.Y, corner2.Y);
+    }
+
+    private static bool PointBetween(Point64 pt, Point64 corner1, Point64 corner2)
+    {
+      // NB points may not be collinear
+      return
+        ValueBetween(pt.X, corner1.X, corner2.X) &&
+        ValueBetween(pt.Y, corner1.Y, corner2.Y);
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -2546,7 +2551,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private Joiner? GetHorzTrialParent(OutPt op)
+    private static Joiner? GetHorzTrialParent(OutPt op)
     {
       Joiner? joiner = op.joiner;
       while (joiner != null)
@@ -2568,7 +2573,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private bool OutPtInTrialHorzList(OutPt op)
+    private static bool OutPtInTrialHorzList(OutPt op)
     {
       return op.joiner != null &&
        ((op.joiner.idx < 0) || GetHorzTrialParent(op) != null);
@@ -2736,7 +2741,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private bool GetHorzExtendedHorzSeg(ref OutPt op, out OutPt op2)
+    private static bool GetHorzExtendedHorzSeg(ref OutPt op, out OutPt op2)
     {
       OutRec outRec = GetRealOutRec(op.outrec)!;
       op2 = op;
@@ -2829,9 +2834,8 @@ namespace Clipper2Lib
         // don't waste time trying to join adjacent vertices
         ((op1.next == op2) && (op1 != op1.outrec.pts)) ||
         ((op2.next == op1) && (op2 != op1.outrec.pts)))) return;
-        
-      Joiner joiner = new Joiner(op1, op2, null);
-      joiner.idx = _joinerList.Count;
+
+      Joiner joiner = new Joiner(op1, op2, null) {idx = _joinerList.Count};
       _joinerList.Add(joiner);
     }
 
@@ -2996,7 +3000,7 @@ namespace Clipper2Lib
             // by inserting an extra vertex if needed
             if (op1.prev.pt != op2.next.pt)
             {
-              if (PointBetween(op1.prev.pt, op2.pt, op2.next.pt))
+              if (PointEqualOrBetween(op1.prev.pt, op2.pt, op2.next.pt))
                 op2.next = InsertOp(op1.prev.pt, op2);
               else
                 op1.prev = InsertOp(op2.next.pt, op1.prev);
@@ -3061,7 +3065,7 @@ namespace Clipper2Lib
             // by inserting an extra vertex if needed
             if (op2.prev.pt != op1.next.pt)
             {
-              if (PointBetween(op2.prev.pt, op1.pt, op1.next.pt))
+              if (PointEqualOrBetween(op2.prev.pt, op1.pt, op1.next.pt))
                 op1.next = InsertOp(op2.prev.pt, op1);
               else
                 op2.prev = InsertOp(op1.next.pt, op2.prev);
@@ -3207,8 +3211,7 @@ namespace Clipper2Lib
 
         if (_using_polytree)
         {
-          if (outrec.splits == null)
-            outrec.splits = new List<OutRec>();
+          outrec.splits ??= new List<OutRec>();
           outrec.splits.Add(newOr);
         }
 
@@ -3339,7 +3342,7 @@ namespace Clipper2Lib
       }
     }
 
-    internal bool BuildPath(OutPt op, bool reverse, bool isOpen, Path64 path)
+    internal static bool BuildPath(OutPt op, bool reverse, bool isOpen, Path64 path)
     {
       if (op.next == op || (!isOpen && op.next == op.prev)) return false;
       path.Clear();
@@ -3404,7 +3407,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private bool Path1InsidePath2(OutRec or1, OutRec or2)
+    private static bool Path1InsidePath2(OutRec or1, OutRec or2)
     {
       PointInPolygonResult result;
       OutPt op = or1.pts!;
@@ -3420,7 +3423,7 @@ namespace Clipper2Lib
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private Rect64 GetBounds(Path64 path)
+    private static Rect64 GetBounds(Path64 path)
 	  {
 		  if (path.Count == 0) return new Rect64();
       Rect64 result = new Rect64(long.MaxValue, long.MaxValue, -long.MaxValue, -long.MaxValue);
@@ -3923,7 +3926,7 @@ namespace Clipper2Lib
     public double Area()
     {
       double result = Polygon == null ? 0 : Clipper.Area(Polygon);
-      foreach (var polyPathBase in _childs)
+      foreach (PolyPathBase polyPathBase in _childs)
       {
         PolyPath64 child = (PolyPath64) polyPathBase;
         result += child.Area();
@@ -3963,7 +3966,7 @@ namespace Clipper2Lib
     public double Area()
     {
       double result = Polygon == null ? 0 : Clipper.Area(Polygon);
-      foreach (var polyPathBase in _childs)
+      foreach (PolyPathBase polyPathBase in _childs)
       {
         PolyPathD child = (PolyPathD) polyPathBase;
         result += child.Area();
