@@ -1,7 +1,7 @@
 ﻿/*******************************************************************************
 * Author    :  Angus Johnson                                                   *
 * Version   :  Clipper2 - ver.1.0.4                                            *
-* Date      :  5 September 2022                                                *
+* Date      :  16 September 2022                                               *
 * Website   :  http://www.angusj.com                                           *
 * Copyright :  Angus Johnson 2010-2022                                         *
 * Purpose   :  This module contains simple functions that will likely cover    *
@@ -23,19 +23,17 @@ namespace Clipper2Lib
   // PRE-COMPILER CONDITIONAL ...
   // USINGZ: For user defined Z-coordinates. See Clipper.SetZ
 
-  using Path64 = List<Point64>;
-  using Paths64 = List<List<Point64>>;
-  using PathD = List<PointD>;
-  using PathsD = List<List<PointD>>;
-
   public static class Clipper
   {
 
-    public static Rect64 MaxInvalidRect64 = new Rect64(
+    private static Rect64 maxInvalidRect64 = new Rect64(
       long.MaxValue, long.MaxValue, long.MinValue, long.MinValue);
 
-    public static RectD MaxInvalidRectD = new RectD(
-      double.MaxValue, -double.MaxValue, -double.MaxValue, -double.MaxValue);
+    private static RectD maxInvalidRectD = new RectD(
+      double.MaxValue, double.MaxValue, -double.MaxValue, -double.MaxValue);
+
+    public static Rect64 MaxInvalidRect64 => maxInvalidRect64;
+    public static RectD MaxInvalidRectD => maxInvalidRectD;
 
     public static Paths64 Intersect(Paths64 subject, Paths64 clip, FillRule fillRule)
     {
@@ -212,28 +210,28 @@ namespace Clipper2Lib
     {
       string result = "";
       foreach (Point64 pt in path)
-        result = result + pt.ToString();
+        result += pt.ToString();
       return result + '\n';
     }
     public static string Paths64ToString(Paths64 paths)
     {
       string result = "";
       foreach (Path64 path in paths)
-        result = result + Path64ToString(path);
+        result += Path64ToString(path);
       return result;
     }
     public static string PathDToString(PathD path)
     {
       string result = "";
       foreach (PointD pt in path)
-        result = result + pt.ToString();
+        result += pt.ToString();
       return result + '\n';
     }
     public static string PathsDToString(PathsD paths)
     {
       string result = "";
       foreach (PathD path in paths)
-        result = result + PathDToString(path);
+        result += PathDToString(path);
       return result;
     }
     public static Path64 OffsetPath(Path64 path, long dx, long dy)
@@ -581,7 +579,7 @@ namespace Clipper2Lib
     public static PathsD PolyTreeToPathsD(PolyTreeD polyTree)
     {
       PathsD result = new PathsD();
-      foreach (var polyPathBase in polyTree)
+      foreach (PolyPathBase polyPathBase in polyTree)
       {
         PolyPathD p = (PolyPathD)polyPathBase;
         AddPolyNodeToPathsD(p, result);
@@ -726,7 +724,7 @@ namespace Clipper2Lib
       else
       {
         while (result.Count > 2 && InternalClipper.CrossProduct(
-          result[result.Count - 1], result[result.Count - 2], result[0]) == 0)
+          result[^1], result[^2], result[0]) == 0)
             result.RemoveAt(result.Count - 1);
         if (result.Count < 3)
           result.Clear();
@@ -749,5 +747,49 @@ namespace Clipper2Lib
       return InternalClipper.PointInPolygon(pt, polygon);
     }
 
-  }
+    public static Path64 Ellipse(Point64 center,
+      double radiusX, double radiusY = 0, int steps = 0)
+    {
+      if (radiusX <= 0) return new Path64();
+      if (radiusY <= 0) radiusY = radiusX;
+      if (steps <= 2)
+        steps = (int) Math.Ceiling(Math.PI * Math.Sqrt((radiusX + radiusY) / 2));
+
+      double si = Math.Sin(2 * Math.PI / steps);
+      double co = Math.Cos(2 * Math.PI / steps);
+      double dx = co, dy = si;
+      Path64 result = new Path64(steps) { new Point64(center.X + radiusX, center.Y) };
+      for (int i = 1; i < steps; ++i)
+      {
+        result.Add(new Point64(center.X + radiusX * dx, center.Y + radiusY * dy));
+        double x = dx * co - dy * si;
+        dy = dy * co + dx * si;
+        dx = x;
+      }
+      return result;
+    }
+
+    public static PathD Ellipse(PointD center,
+      double radiusX, double radiusY = 0, int steps = 0)
+    {
+      if (radiusX <= 0) return new PathD();
+      if (radiusY <= 0) radiusY = radiusX;
+      if (steps <= 2)
+        steps = (int) Math.Ceiling(Math.PI * Math.Sqrt((radiusX + radiusY) / 2));
+
+      double si = Math.Sin(2 * Math.PI / steps);
+      double co = Math.Cos(2 * Math.PI / steps);
+      double dx = co, dy = si;
+      PathD result = new PathD(steps) { new PointD(center.x + radiusX, center.y) };
+      for (int i = 1; i < steps; ++i)
+      {
+        result.Add(new PointD(center.x + radiusX * dx, center.y + radiusY * dy));
+        double x = dx * co - dy * si;
+        dy = dy * co + dx * si;
+        dx = x;
+      }
+      return result;
+    }
+
+  } // Clipper
 } // namespace
