@@ -1,21 +1,15 @@
 ﻿/*******************************************************************************
 * Author    :  Angus Johnson                                                   *
-* Date      :  3 July 2022                                                     *
+* Date      :  16 September 2022                                               *
 * Website   :  http://www.angusj.com                                           *
 * Copyright :  Angus Johnson 2010-2022                                         *
-* License:                                                                     *
-* Use, modification & distribution is subject to Boost Software License Ver 1. *
-* http://www.boost.org/LICENSE_1_0.txt                                         *
+* License   :  http://www.boost.org/LICENSE_1_0.txt                            *
 *******************************************************************************/
 
-using System.Collections.Generic;
 using System.IO;
 
 namespace Clipper2Lib
 {
-
-  using Paths64 = List<List<Point64>>;
-  using PathsD = List<List<PointD>>;
 
   public static class SvgUtils
   {
@@ -34,7 +28,21 @@ namespace Clipper2Lib
         svg.AddPaths(paths, false, 0x1800009C, 0xCCB3B3DA, 0.8);
     }
 
+    public static void AddSubject(SimpleSvgWriter svg, PathsD paths,
+      bool is_closed = true, bool is_joined = true)
+    {
+      if (!is_closed)
+        svg.AddPaths(paths, !is_joined, 0x0, 0xCCB3B3DA, 0.8);
+      else
+        svg.AddPaths(paths, false, 0x1800009C, 0xCCB3B3DA, 0.8);
+    }
+
     public static void AddClip(SimpleSvgWriter svg, Paths64 paths)
+    {
+      svg.AddPaths(paths, false, 0x129C0000, 0xCCFFA07A, 0.8);
+    }
+
+    public static void AddClip(SimpleSvgWriter svg, PathsD paths)
     {
       svg.AddPaths(paths, false, 0x129C0000, 0xCCFFA07A, 0.8);
     }
@@ -45,7 +53,7 @@ namespace Clipper2Lib
       if (!is_closed)
         svg.AddPaths(paths, !is_joined, 0x0, 0xFF003300, 0.8, show_coords);
       else
-        svg.AddPaths(paths, false, 0xFF80ff9C, 0xFF003300, 0.8, show_coords);
+        svg.AddPaths(paths, false, 0x4080ff9C, 0xFF003300, 0.8, show_coords);
     }
 
     public static void AddSolution(SimpleSvgWriter svg, PathsD paths,
@@ -54,7 +62,7 @@ namespace Clipper2Lib
       if (!is_closed)
         svg.AddPaths(paths, !is_joined, 0x0, 0xFF003300, 0.8, show_coords);
       else
-        svg.AddPaths(paths, false, 0xFF80ff9C, 0xFF003300, 0.8, show_coords);
+        svg.AddPaths(paths, false, 0x8080ff9C, 0xFF003300, 0.8, show_coords);
     }
 
     public static void SaveToFile(SimpleSvgWriter svg,
